requirements.txt
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.10.3
cycler==0.12.1
dash==3.2.0
Django==5.2.5
django-auditlog==3.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
factory_boy==3.3.3
Faker==37.5.3
Flask==3.1.1
fonttools==4.59.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
narwhals==2.1.0
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
pluggy==1.6.0
psycopg2-binary==2.9.10
py4j==0.10.9.9
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pyspark==4.0.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retrying==1.4.2
rpds-py==0.27.0
seaborn==0.13.2
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
zipp==3.23.0
